refactor: Package 명 adpater로 변경
