+++ event-service/src/main/java/com/example/eventservice/domain/adapter/interaction/InteractionRepository.java
@@ -1,4 +1,4 @@
-package com.example.eventservice.domain.repository.interaction;
+package com.example.eventservice.domain.adapter.interaction;
 
 import com.example.eventservice.domain.entity.interaction.LikeStar;
 import com.example.eventservice.domain.entity.interaction.Interaction;
