+++ event-service/src/main/java/com/example/eventservice/domain/adapter/interaction/query/WhereQuery.java
@@ -1,4 +1,4 @@
-package com.example.eventservice.domain.repository.interaction.query;
+package com.example.eventservice.domain.adapter.interaction.query;
 
 import com.example.eventservice.domain.entity.interaction.LikeStar;
 import com.querydsl.core.types.dsl.BooleanExpression;
